backend/requirements.txt
networkx>=3.1
pandas>=2.0
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic[email]
python-dotenv==1.0.1
requests==2.32.3
aiohttp==3.11.12
scikit-learn==1.6.1
numpy==1.26.4
web3==7.6.0
sqlalchemy==2.0.36
psycopg2-binary>=2.9.0
PyJWT==2.8.0
reportlab==4.2.5
slowapi==0.1.9
bcrypt>=4.0.0
stripe>=7.0.0
email-validator>=2.0.0
